+++ src/host/plugin/au_plugin.h
@@ -75,6 +75,6 @@ public:
     void loadSnapshot(const plugin_snapshot_t& pluginSnapshot) override;
     guiplugin* makeGui() override;
     guiplugin* getGui() override;
-    void process(AudioBlock* in, AudioBlock* out, double tick, int32_t samplePos, int32_t numSamples, playback_state state) override;
+    void process(AudioBlock* in, AudioBlock* out, double tick, double samplePos, int32_t numSamples, playback_state state) override;
     int32_t getPluginLatency() override;
 };
