+++ src/host/plugin/base_plugin.h
@@ -107,7 +107,7 @@ public:
     virtual guiplugin* getGui()  = 0;
 
     virtual void makeSnapshot(plugin_snapshot_t& ps, const tracksnapshot_store_opts_t& opts) = 0;
-    virtual void process(AudioBlock* in, AudioBlock* out, double tick, int32_t samplePos, int32_t numSamples, playback_state state) = 0;
+    virtual void process(AudioBlock* in, AudioBlock* out, double tick, double samplePos, int32_t numSamples, playback_state state) = 0;
     virtual void postProcess(AudioBlock* out, int32_t samples, bool hasProcessed);
     virtual bool show()   = 0;
     virtual bool close()  = 0;
@@ -187,7 +187,7 @@ public:
     void makeSnapshot(plugin_snapshot_t& ps, const tracksnapshot_store_opts_t& opts) override;
     guiplugin* makeGui() override;
     guiplugin* getGui() override;
-    void process(AudioBlock* in, AudioBlock* out, double tick, int32_t samplePos, int32_t numSamples, playback_state state) override;
+    void process(AudioBlock* in, AudioBlock* out, double tick, double samplePos, int32_t numSamples, playback_state state) override;
     bool show() override;
     bool close() override;
