+++ src/host/plugin/empty.cpp
@@ -70,7 +70,7 @@ int32_t module_empty::getPluginLatency() {
     return 0;
 }
 
-void module_empty::process(AudioBlock* in, AudioBlock* out, double tick, int32_t samplePos, int32_t numSamples, playback_state state) {
+void module_empty::process(AudioBlock* in, AudioBlock* out, double tick, double samplePos, int32_t numSamples, playback_state state) {
     dbgassert(getTrackLink()->sampleFormat == this->format
               && in->samples == format.blockSize
               && out->samples == format.blockSize
