+++ src/host/plugin/group.h
@@ -26,7 +26,7 @@ public:
     guiplugin* makeGui() override;
     guiplugin* getGui() override;
     int32_t getPluginLatency() override;
-    void process(AudioBlock* in, AudioBlock* out, double tick, int32_t samplePos, int32_t numSamples, playback_state state) override;
+    void process(AudioBlock* in, AudioBlock* out, double tick, double samplePos, int32_t numSamples, playback_state state) override;
     String getInfo(std::vector<String>& list) override;
     void onEnable() override;
     void onDisable() override;
