+++ src/host/plugin/vst_plugin.cpp
@@ -846,7 +846,7 @@ int64_t vstplugin::dispatch(
         float opt) {
     return vst_dispatch(this, this->handle->aeffect, opcode, index, value, ptr, opt);
 }
-void vstplugin::process(AudioBlock* in, AudioBlock* out, double tick, int32_t samplePos, int32_t numSamples, playback_state state) {
+void vstplugin::process(AudioBlock* in, AudioBlock* out, double tick, double samplePos, int32_t numSamples, playback_state state) {
     dbgassert(!isInSuspend);
     dbgassert(this->handle->aeffect);
     dbgassert(getTrackLink()->sampleFormat == this->format && in->samples == format.blockSize && out->samples == format.blockSize && format.blockSize > 0 && format.sampleRate > 0);
