+++ src/host/plugin/vst_plugin.h
@@ -137,7 +137,7 @@ public:
     void loadSnapshot(const plugin_snapshot_t& pluginSnapshot) override;
     guiplugin* makeGui() override;
     guiplugin* getGui() override;
-    void process(AudioBlock* in, AudioBlock* out, double tick, int32_t samplePos, int32_t numSamples, playback_state state) override;
+    void process(AudioBlock* in, AudioBlock* out, double tick, double samplePos, int32_t numSamples, playback_state state) override;
     int32_t getPluginLatency() override;
     int32_t getFlagsVST();
     VstTimeInfo* getLocalTimeInfoPtr();
